requirements.txt
selenium==3.141.0
msedge-selenium-tools==3.141.3